backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.36
aiopg==1.4.0
alembic==1.14.0
pydantic==2.10.0
pydantic-settings==2.6.0
httpx==0.25.2
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
python-multipart==0.0.6